coalals.results: Refactor to submodule
To make room for the new additions to results
module, this is being refactored to a new sub
module.
